requirements.txt
requests
beautifulsoup4
python-dateutil
psycopg2-binary
python-dotenv
